python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
aiosqlite==0.19.0
opencv-python==4.8.1.78
scikit-learn-intelex==2024.0.1
numpy==1.24.3
pillow==10.1.0
pydantic==2.4.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-dotenv==1.0.0
websockets==12.0
reportlab==4.1.0
python-dateutil==2.9.0